compute_node/requirements.txt
PySide6
pygame
pyzmq
numpy
opencv-python
ultralytics